Googletest export
Replace pump'd Args() matcher with variadic templates.

PiperOrigin-RevId: 220847745
